Title: 시력 검사 2, Time: 108ms, Memory: 24MB, Status: Passed - CodeTree

+++ 240812/시력 검사 2/eye-test-2.py
@@ -1,4 +1,4 @@
-a = int(input())
+a = float(input())
 
 if a >= 1 :
     print("High")
